+++ backends/optype.go
@@ -151,7 +151,7 @@ const (
 	OpTypeFusedScaledDotProductAttention
 	OpTypeFusedAttentionQKVProjection
 	OpTypeFusedQuantizedDense
-	OpTypeFusedQuantizedGather
+	OpTypeQuantizedEmbeddingLookup
 
 	// OpTypeLast should always be kept the last, it is used as a counter/marker for OpType.
 	OpTypeLast